+++ docs/app/documentation/utilities/api-files.ts
@@ -178,7 +178,7 @@ export const API_FILES = {
         'SideNavigationTitleComponent'
     ],
     table: [
-        'TableComponent',
+        'TableDirective',
         'ColumnSortableDirective'
     ],
     tabs: [