+++ library/src/lib/calendar/calendar.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IconModule } from '../icon/icon.module';
-import { UtilsModule } from '../utils/utils.module';
 
 import { CalendarComponent } from './calendar.component';
 
 @NgModule({
     declarations: [CalendarComponent],
-    imports: [CommonModule, IconModule, UtilsModule],
+    imports: [CommonModule, IconModule],
     exports: [CalendarComponent]
 })
 export class CalendarModule {}