+++ CHANGELOG.md
@@ -2,6 +2,66 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.5.2"></a>
+## [0.5.2](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.6...v0.5.2) (2019-05-30)
+
+
+
+<a name="0.5.2-rc.6"></a>
+## [0.5.2-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.5...v0.5.2-rc.6) (2019-05-29)
+
+
+### Bug Fixes
+
+* remove utilsmodule, hashservice, refactor unique id approach ([#847](https://github.com/SAP/fundamental-ngx/issues/847)) ([ec9abc7](https://github.com/SAP/fundamental-ngx/commit/ec9abc7))
+
+
+
+<a name="0.5.2-rc.5"></a>
+## [0.5.2-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.4...v0.5.2-rc.5) (2019-05-29)
+
+
+### Bug Fixes
+
+* allow datetime null ([#837](https://github.com/SAP/fundamental-ngx/issues/837)) ([baee9d9](https://github.com/SAP/fundamental-ngx/commit/baee9d9))
+
+
+
+<a name="0.5.2-rc.4"></a>
+## [0.5.2-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.3...v0.5.2-rc.4) (2019-05-29)
+
+
+### Bug Fixes
+
+* dismissing alerts display: none style fix ([#848](https://github.com/SAP/fundamental-ngx/issues/848)) ([45a8dbe](https://github.com/SAP/fundamental-ngx/commit/45a8dbe))
+
+
+
+<a name="0.5.2-rc.3"></a>
+## [0.5.2-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.2...v0.5.2-rc.3) (2019-05-27)
+
+
+
+<a name="0.5.2-rc.2"></a>
+## [0.5.2-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.1...v0.5.2-rc.2) (2019-05-27)
+
+
+
+<a name="0.5.2-rc.1"></a>
+## [0.5.2-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.5.2-rc.0...v0.5.2-rc.1) (2019-05-24)
+
+
+
+<a name="0.5.2-rc.0"></a>
+## [0.5.2-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.5.1...v0.5.2-rc.0) (2019-05-23)
+
+
+
+<a name="0.5.1"></a>
+## [0.5.1](https://github.com/SAP/fundamental-ngx/compare/v0.5.1-rc.10...v0.5.1) (2019-05-22)
+
+
+
 <a name="0.5.1-rc.10"></a>
 ## [0.5.1-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.5.1-rc.9...v0.5.1-rc.10) (2019-05-19)